Also move the copy string functions into ember strings. header af.h and implementation in util.cpp makes no sense
