+++ src/app/util/ember-strings.cpp
@@ -34,3 +34,50 @@ uint16_t emberAfLongStringLength(const uint8_t * buffer)
     uint16_t length = Encoding::LittleEndian::Get16(buffer);
     return (length == 0xFFFF ? 0 : length);
 }
+
+void emberAfCopyString(uint8_t * dest, const uint8_t * src, size_t size)
+{
+    if (src == nullptr)
+    {
+        dest[0] = 0; // Zero out the length of string
+    }
+    else if (src[0] == 0xFF)
+    {
+        dest[0] = src[0];
+    }
+    else
+    {
+        uint8_t length = emberAfStringLength(src);
+        if (size < length)
+        {
+            // Since we have checked that size < length, size must be able to fit into the type of length.
+            length = static_cast<decltype(length)>(size);
+        }
+        memmove(dest + 1, src + 1, length);
+        dest[0] = length;
+    }
+}
+
+void emberAfCopyLongString(uint8_t * dest, const uint8_t * src, size_t size)
+{
+    if (src == nullptr)
+    {
+        dest[0] = dest[1] = 0; // Zero out the length of string
+    }
+    else if ((src[0] == 0xFF) && (src[1] == 0xFF))
+    {
+        dest[0] = 0xFF;
+        dest[1] = 0xFF;
+    }
+    else
+    {
+        uint16_t length = emberAfLongStringLength(src);
+        if (size < length)
+        {
+            // Since we have checked that size < length, size must be able to fit into the type of length.
+            length = static_cast<decltype(length)>(size);
+        }
+        memmove(dest + 2, src + 2, length);
+        Encoding::LittleEndian::Put16(dest, length);
+    }
+}
